Refactor: setErrorMessage
