+++ assets/js/core/core-form.js
@@ -1,3 +1,5 @@
+import { showError } from "../utils/ui.js"
+import { errorMessages } from "../utils/validation.js"
 import { formState } from "./index.js"
 
 export default class Form {
@@ -19,6 +21,19 @@ export default class Form {
     for (const element of this.inputElements) {
       formState.data = { name: element.name, value: element.value }
     }
+  }
+
+  setErrorMessage(inputElement, errorType) {
+    const inputRootElement = document.querySelector(`.input-layout-${inputElement.name}`)
+    const errorElement = inputRootElement.querySelector(".input-error")
+    const errorMessage = errorMessages[inputElement.name][errorType]
+
+    return showError({
+      inputRootElement,
+      errorElement,
+      errorMessage: errorMessage,
+      className: this.ERROR_CLASS_NAME,
+    })
   }
 
   attach() {
